backend/requirements.txt
flask
flask-cors
pyodbc
python-dotenv
pandas
sqlalchemy
psycopg2-binary
google-api-python-client
google-auth-httplib2
google-auth-oauthlib